+++ packages/solidart/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.5.0
+
+- **FEAT**: Add `run` method to `Computed` to manually trigger an update of its value.
+- **FEAT**: Add `run` method to `Effect` to manually re-run the effect.
+- **CHORE**: Detect if `Effect` didn't track any reactive value and throw an `EffectWithoutDependenciesError`.
+- **CHORE**: Detect if `SignalBuilder` didn't track any reactive value and throw a `SignalBuilderWithoutDependenciesError`.
+
 ## 2.4.1
 
 - **FIX**: `Signal.lazy` which caused an exception.